Add shovel resource and data source
Implements lavinmq_shovel resource and lavinmq_shovels data source
for configuring message forwarding between queues and exchanges.

Decided to create a specific shovel resource instead of a "generic"
parameters resource. I think this is easier to implement (schema and
validation is more clear) and easier to understand as an end user.